Refactor: Move `core.preview` components to `core.ui.preview` and update imports across the codebase.
